Product variant attribute and variant detail have been covered by unit tests

+++ tests/RenStore.Catalog.Tests/UnitTets/Persistence/Repositories/CategoryRepository/AddAsyncTests.cs
@@ -4,14 +4,23 @@ using RenStore.Catalog.Application.Abstractions;
 using RenStore.Catalog.Domain.Aggregates.Category;
 using RenStore.Catalog.Persistence;
 
-namespace RenStore.Catalog.Tests.UnitTets.Domain.Repositories.CategoryRepository;
+namespace RenStore.Catalog.Tests.UnitTets.Persistence.Repositories.CategoryRepository;
 
 public class AddAsyncTests : IAsyncLifetime
 {
-    private static string _connectionString =
-        $"Server=localhost;Port=5432;DataBase={Guid.NewGuid()}; User Id=re;Password=postgres;Include Error Detail=True";
-    
     private CatalogDbContext _context;
+    
+    public async Task InitializeAsync()
+    {
+        var options = new DbContextOptionsBuilder<CatalogDbContext>()
+            .UseNpgsql(connectionString: CatalogRepositoryTestsBase
+                .BuildConnectionString(Guid.NewGuid()))
+            .Options;
+
+        _context = new CatalogDbContext(options);
+        await _context.Database.EnsureDeletedAsync();
+        await _context.Database.EnsureCreatedAsync();
+    }
     
     [Fact]
     public async Task Should_Save_Category_To_Postgres()
@@ -21,19 +30,10 @@ public class AddAsyncTests : IAsyncLifetime
         var categoryNameRu = "Одежда";
         var description = "Sample testing category description";
         var now = DateTimeOffset.UtcNow;
-        
-        var options = new DbContextOptionsBuilder<CatalogDbContext>()
-            .UseNpgsql(connectionString: _connectionString)
-            .Options;
-
-        _context = new CatalogDbContext(options);
-        await _context.Database.EnsureDeletedAsync();
-        await _context.Database.EnsureCreatedAsync();
-        /*await context.Database.MigrateAsync();*/
 
         var eventStoreMock = new Mock<IEventStore>();
         
-        var repository = new Persistence.Write.Repositories.Postgresql
+        var repository = new Catalog.Persistence.Write.Repositories.Postgresql
             .CategoryRepository(
                 _context, 
                 eventStoreMock.Object);
@@ -64,18 +64,10 @@ public class AddAsyncTests : IAsyncLifetime
     {
         // Arrange
         Category category = null;
-        
-        var options = new DbContextOptionsBuilder<CatalogDbContext>()
-            .UseNpgsql(connectionString: _connectionString)
-            .Options;
-
-        _context = new CatalogDbContext(options);
-        await _context.Database.EnsureDeletedAsync();
-        await _context.Database.EnsureCreatedAsync();
 
         var eventStoreMock = new Mock<IEventStore>();
         
-        var repository = new Persistence.Write.Repositories.Postgresql
+        var repository = new Catalog.Persistence.Write.Repositories.Postgresql
             .CategoryRepository(
                 _context, 
                 eventStoreMock.Object);
@@ -86,13 +78,10 @@ public class AddAsyncTests : IAsyncLifetime
                 category: category!, 
                 CancellationToken.None));
     }
-    
-    public async Task InitializeAsync()
-    {
-    }
 
     public async Task DisposeAsync()
     {
-        await _context.Database.EnsureDeletedAsync();
+        if(_context != null)
+         await _context.Database.EnsureDeletedAsync();
     }
 }
